Add separate workspaces for main and staging, and separate scripts for main and staging environments.
