+++ infra/.gitignore
@@ -1,5 +1,6 @@
 tsconfig.tsbuildinfo
 cdktf.out/
+dist/
 
 # This package does not build to a dist directory, so add explicity exclude
 # patterns for the Typescript artifacts.
