+++ infra/package.json
@@ -6,10 +6,13 @@
   "scripts": {
     "build": "pnpm build:tsc && pnpm build:synth",
     "build:get": "cdktf get",
-    "build:synth": "cdktf synth",
+    "build:synth": "pnpm build:synth:main && pnpm build:synth:staging",
+    "build:synth:main": "cdktf synth --app 'npx ts-node src/spaces/main.ts'",
+    "build:synth:staging": "cdktf synth --app 'npx ts-node src/spaces/staging.ts'",
     "build:tsc": "tsc --pretty",
     "clean": "rm -rf .gen cdktf.out",
-    "deploy": "cdktf deploy",
+    "deploy:main": "cdktf deploy --app 'npx ts-node src/spaces/main.ts'",
+    "deploy:staging": "cdktf deploy --app 'npx ts-node src/spaces/staging.ts'",
     "dev": "tsc -w",
     "test": "jest",
     "test:watch": "jest --watch"
