--- infra/src/index.ts
@@ -1,3 +1,0 @@
-import { registerAppStack } from './app-stack';
-
-registerAppStack('tts-10x-atj-dev');
