+++ infra/src/spaces/main.ts
@@ -1,0 +1,3 @@
+import { registerAppStack } from '../lib/app-stack';
+
+registerAppStack('tts-10x-atj-dev');
